WIP: adding logic to when messages are shown
